services/notification-service/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
gunicorn==21.2.0
pika==1.3.2
python-dotenv==1.0.0
dj-database-url==2.1.0
PyJWT==2.8.0
celery==5.3.6
